Added ability to change incompatible encoders

+++ HandBrakeCLIBatchEncode/Business/Processes/Encoder.cs
@@ -21,6 +21,14 @@ namespace HandBrakeCLIBatchEncode
 
         internal static bool ClosingPrematurely { get; set; } = false;
 
+        internal static bool StartEncodeSuccess { get; set; } = false;
+
+        internal static bool TryAnotherEncoder { get; set; } = false;
+
+        internal static int EncoderAttempt { get; set; } = 1;
+
+        internal static bool DontDeleteTempFile { get; set; } = false;
+
         public void EncodeVideos(string rootFileOrCombined, string presetPath, string presetName, string audioByteRate)
         {
             handler = new ConsoleEventDelegate(ConsoleEventCallback);
@@ -51,6 +59,11 @@ namespace HandBrakeCLIBatchEncode
                 {
                     if (ClosingPrematurely)
                         break;
+
+                    TryAnotherEncoder = false;
+                    StartEncodeSuccess = false;
+                    DontDeleteTempFile = false;
+                    EncoderAttempt = 1;
 
                     FileInfo info = new FileInfo(file);
 
@@ -154,6 +167,35 @@ namespace HandBrakeCLIBatchEncode
                 process.BeginOutputReadLine();
 
                 process.WaitForExit();
+
+                if (TryAnotherEncoder)
+                {
+                    if (EncoderAttempt == 1)
+                        PresetValidator.ChangeEncoder(presetPath, "nvenc_h265", "qsv_h264");
+                    else if (EncoderAttempt == 2)
+                        PresetValidator.ChangeEncoder(presetPath, "qsv_h264", "nvenc_h265");
+                    else if (EncoderAttempt == 3)
+                        PresetValidator.ChangeEncoder(presetPath, "qsv_h264", "x264"); // software - last resort - should always work
+                    else if (EncoderAttempt == 4)
+                        PresetValidator.ChangeEncoder(presetPath, "qsv_h264", "x264"); // software - last resort - should always work
+                    else if (EncoderAttempt == 5)
+                        PresetValidator.ChangeEncoder(presetPath, "x264", "qsv_h264");
+                    else if (EncoderAttempt == 6)
+                        PresetValidator.ChangeEncoder(presetPath, "x264", "nvenc_h265");
+                    else if (EncoderAttempt > 6)
+                    {
+                        TryAnotherEncoder = false;
+                        DontDeleteTempFile = true;
+
+                        Console.ForegroundColor = ConsoleColor.Red;
+                        WriteAndRecord(" FAIL");
+                        Console.ResetColor();
+                    }
+
+                    EncoderAttempt++;
+
+                    return PerformVideoEncode(inputFile, outputFile, presetPath, presetName, audioByteRate);
+                }
             }
 
             #region Delete temp file
@@ -161,32 +203,35 @@ namespace HandBrakeCLIBatchEncode
             // Need to delay so that if closing the application prematurely, run the 'ConsoleEventCallback' method first so that
             // We don't delete the original file
 
-            new Thread((ThreadStart)delegate
+            if (!DontDeleteTempFile)
             {
-                Thread.Sleep(2000);
-
-                if (!ClosingPrematurely)
+                new Thread((ThreadStart)delegate
                 {
-                    try
-                    {
-                        File.Delete(inputFile);
-                    }
-                    catch
-                    {
-                        Thread.Sleep(500);
+                    Thread.Sleep(2000);
 
+                    if (!ClosingPrematurely)
+                    {
                         try
                         {
                             File.Delete(inputFile);
                         }
                         catch
                         {
-                            // Forget it
+                            Thread.Sleep(500);
+
+                            try
+                            {
+                                File.Delete(inputFile);
+                            }
+                            catch
+                            {
+                                // Forget it
+                            }
                         }
                     }
-                }
 
-            }).Start();
+                }).Start();
+            }
 
             #endregion
